updated lootTables Formula and added logic to bulletin boards
LootTable was only adding 1-4 levels to playerlevel for unleveled items, it should have been adding 1-25 corrected logic
Added logic that when you click on a buletin board in towns, it will update your map to reveal all locations.  The idea being that the bulletin board includes a map of the town.