PlayerEnterExit: Reduce delay filling Dungeon var
Previously, GameObjectHelper ran the SetDungeon function in DaggerfallDungeon before returning it's results. Since PlayerEnterExit uses the value of this returned result to fill it's Dungeon variable, this meant that there was a short period where Dungeon was equal to Null even though the DaggerfallDungeon object existed. This moves the SetDungeon function so that it runs after the Dungeon variable is filled.