Automap: Move Mesh&Mat update to separate script
This moves the contents of the InjectMeshAndMaterialProperties and UpdateMaterialsOfMeshRenderer functions into the AutomapModel script. This script is attached to every model that is used to construct the Automap. Then, instead of the Automap script iterating through the Automap tree to run these function on every model it finds, it raises an event that causes each model to run it's copy of the AutomapModel script functions.
I also made some optimisations to the existing code.