RuntimeMaterials: Use texture table in Automaps
RuntimeMaterials also had a problem where it was using the wrong dungeon texture table to create it's materials.

This time, it was because when a model using it was attached to the Automap, it initalised with the default set of textures then registered itself for the DaggerfallDungeon SetDungeon event to later re-set those textures with the correct texture table. The problem is that DaggerfallDungeon had already called this event by the time the Automap was created and so these event registrations never fired.

Fixed this by making RuntimeMaterials detect if it is being created as part of the Automap. If so, it will immediately generate the materials using the correct texture table instead of waiting for an event.