Added global mod order configuration.  The file is called mod_order.txt in Daggerfall Unity appdata location  If this file does not exist, a template is created.
The format is to use tab delimited
Mod	Override	modName	IsOptional	IsPeer	IsConflict	Version
Mod = the mod to apply the rule
Override is true or false and dictates if you overWrite an existing rule
modName is the dependency for the mod
IsOptional is true or false - does this mod need to exist in the order
IsPeer is true or false, if false, dependent mod must be earlier in the load order
IsConflict is true or false, if true, dependent mod is a conflict to mod; only generates an on screen warning, and message in the log.
Version is optional but if it exists, it checks for a minimum version for the dependent mod

$Disable Conflicts = false
add $disable Conflicts and set to true to automatically disable any dependent mods that are flagged true in IsConflict and are currently enabled.