Added logic to RegisterCustomUIWindow function to honor load order when registering UI Window overrides.
In order to avoid having each mod changed to send mod class into the function, the new function used stacktrace variables and data manipulation to work back to the title of the mod.  This works because most mods use mod.Title as the name of the GameObject.
There was one mod that did not have that logic and hardcoded the name.  In doing so, they also named it differently than Title so it would fail the test.